management-dashboard/src/app/shared/pipe/search-employee.pipe.ts
import { Pipe, PipeTransform } from '@angular/core';
import { Employee } from '../model/employee.model';

@Pipe({
  name: 'searchEmployee',
  standalone: true, // Mark the pipe as standalone
})
export class SearchEmployeePipe implements PipeTransform {
  transform(employees: Employee[], searchTerm: string): Employee[] {
    if (!employees || !searchTerm) {
      return employees;  // If no employees or search term, return the original list
    }

    return employees.filter(employee =>
      employee.firstName.toLowerCase().includes(searchTerm.toLowerCase()) ||
      employee.lastName.toLowerCase().includes(searchTerm.toLowerCase())
    );
  }
  }